+++ constant/config.go
@@ -64,6 +64,7 @@ const (
 	TUIReadOnEnter        = "tui.read_on_enter"
 	TUISearchPromptString = "tui.search_prompt"
 	TUIShowURLs           = "tui.show_urls"
+	TUIShowDownloadedPath = "tui.show_downloaded_path"
 	TUIReverseChapters    = "tui.reverse_chapters"
 )